refactor(test_auth): add conftest.py for fixtures
Now it's possible to use fixtures in other files without
code duplicates
